+++ demo/api/src/auth/static-users.ts
@@ -6,11 +6,13 @@ export const staticUsers: User[] = [
         name: "Admin",
         email: "demo@comet-dxp.com",
         locale: "en",
+        isAdmin: true,
     },
     {
         id: "2",
         name: "Non-Admin",
         email: "test@test.com",
         locale: "en",
+        isAdmin: false,
     },
 ];