+++ demo/api/src/pages/blocks/PageContentBlock.ts
@@ -2,6 +2,7 @@ import { BaseBlocksBlockItemData, BaseBlocksBlockItemInput, BlockField, createBl
 import { AnchorBlock, DamImageBlock, DamVideoBlock } from "@comet/cms-api";
 import { LinkListBlock } from "@src/common/blocks/link-list.block";
 import { RichTextBlock } from "@src/common/blocks/rich-text.block";
+import { NewsDetailBlock } from "@src/news/blocks/news-detail.block";
 import { UserGroup } from "@src/user-groups/user-group";
 import { IsEnum } from "class-validator";
 
@@ -30,6 +31,7 @@ const supportedBlocks = {
     twoLists: TwoListsBlock,
     media: MediaBlock,
     teaser: TeaserBlock,
+    newsDetail: NewsDetailBlock,
 };
 
 class BlocksBlockItemData extends BaseBlocksBlockItemData(supportedBlocks) {