+++ packages/admin/admin-theme/src/paletteOptions.ts
@@ -1,26 +1,55 @@
 import { PaletteOptions } from "@mui/material/styles";
 
-import { bluePalette, errorPalette, greenPalette, infoPalette, neutrals, successPalette, warningPalette } from "./colors";
+import { errorPalette, greyPalette, infoPalette, primaryPalette, secondaryPalette, successPalette, warningPalette } from "./colors";
 
 export const paletteOptions: PaletteOptions = {
-    primary: bluePalette,
-    secondary: greenPalette,
+    primary: primaryPalette,
+    secondary: secondaryPalette,
     error: errorPalette,
-    success: successPalette,
+    warning: warningPalette,
     info: infoPalette,
-    grey: neutrals,
-    divider: neutrals[100],
+    success: successPalette,
+    grey: greyPalette,
+    divider: greyPalette[100],
     text: {
-        primary: neutrals[900],
-        secondary: "#757575",
-        disabled: neutrals[300],
+        primary: greyPalette[900],
+        secondary: greyPalette[300],
+        disabled: greyPalette[200],
     },
     background: {
-        default: neutrals[50],
+        default: greyPalette[50],
     },
     action: {
-        active: neutrals[400],
-        disabled: neutrals[200],
+        active: greyPalette[400],
+        disabled: greyPalette[200],
     },
-    warning: warningPalette,
+    highlight: {
+        purple: "#952F80",
+        green: "#80B50C",
+        orange: "#FFB31A",
+        yellow: "#FFEB3B",
+        red: "#D11700",
+    },
 };
+
+declare module "@mui/material/styles" {
+    interface Palette {
+        highlight: {
+            purple: string;
+            green: string;
+            orange: string;
+            yellow: string;
+            red: string;
+        };
+    }
+
+    interface PaletteOptions {
+        highlight?: {
+            purple: string;
+            green: string;
+            orange: string;
+            yellow: string;
+            red: string;
+        };
+    }
+}