+++ packages/admin/admin-theme/src/typographyOptions.ts
@@ -1,3 +1,4 @@
+import { Breakpoints } from "@mui/material";
 import { TypographyOptions } from "@mui/material/styles/createTypography";
 
 const fontFamily = "Roboto, Helvetica, Arial, sans-serif";
@@ -10,7 +11,8 @@ export const fontWeights = {
     fontWeightBold: 500,
 };
 
-export const typographyOptions: TypographyOptions = {
+export const createTypographyOptions = (breakpoints: Breakpoints): TypographyOptions => ({
+    fontFamily,
     ...fontWeights,
     h1: {
         fontFamily,
@@ -18,7 +20,7 @@ export const typographyOptions: TypographyOptions = {
         lineHeight: "42px",
         fontWeight: fontWeights.fontWeightLight,
 
-        "@media (min-width: 900px)": {
+        [breakpoints.up("md")]: {
             fontSize: 55,
             lineHeight: "64px",
         },
@@ -29,7 +31,7 @@ export const typographyOptions: TypographyOptions = {
         lineHeight: "38px",
         fontWeight: fontWeights.fontWeightSemiLight,
 
-        "@media (min-width: 900px)": {
+        [breakpoints.up("md")]: {
             fontSize: 44,
             lineHeight: "52px",
             fontWeight: fontWeights.fontWeightLight,
@@ -41,7 +43,7 @@ export const typographyOptions: TypographyOptions = {
         lineHeight: "28px",
         fontWeight: fontWeights.fontWeightRegular,
 
-        "@media (min-width: 900px)": {
+        [breakpoints.up("md")]: {
             fontSize: 33,
             lineHeight: "39px",
             fontWeight: fontWeights.fontWeightLight,
@@ -53,7 +55,7 @@ export const typographyOptions: TypographyOptions = {
         lineHeight: "26px",
         fontWeight: fontWeights.fontWeightRegular,
 
-        "@media (min-width: 900px)": {
+        [breakpoints.up("md")]: {
             fontSize: 24,
             lineHeight: "28px",
             fontWeight: fontWeights.fontWeightSemiLight,
@@ -65,7 +67,7 @@ export const typographyOptions: TypographyOptions = {
         lineHeight: "20px",
         fontWeight: fontWeights.fontWeightMedium,
 
-        "@media (min-width: 900px)": {
+        [breakpoints.up("md")]: {
             fontSize: 18,
             lineHeight: "21px",
             fontWeight: fontWeights.fontWeightRegular,
@@ -77,7 +79,7 @@ export const typographyOptions: TypographyOptions = {
         lineHeight: "18px",
         fontWeight: fontWeights.fontWeightBold,
 
-        "@media (min-width: 900px)": {
+        [breakpoints.up("md")]: {
             fontSize: 16,
             lineHeight: "20px",
         },
@@ -94,4 +96,4 @@ export const typographyOptions: TypographyOptions = {
         lineHeight: "20px",
         fontWeight: fontWeights.fontWeightRegular,
     },
-};
+});