+++ packages/admin/admin/src/tabs/Tabs.tsx
@@ -32,7 +32,7 @@ const Content = createComponentSlot("div")<TabsClassKey, OwnerState>({
     ({ ownerState }) => css`
         ${ownerState.contentHidden &&
         css`
-            display: "none";
+            display: none;
         `}
     `,
 );