+++ packages/admin/cms-admin/src/generator/future/generator.ts
@@ -66,6 +66,7 @@ export type GridConfig<T extends { __typename?: string }> = {
     delete?: boolean;
     copyPaste?: boolean;
     readOnly?: boolean;
+    filterProp?: boolean;
 };
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any