+++ packages/api/cms-api/src/user-permissions/dto/user.ts
@@ -1,7 +1,9 @@
 import { Field, ObjectType } from "@nestjs/graphql";
 
+import { User as UserInterface } from "../interfaces/user";
+
 @ObjectType()
-export class User {
+export class User implements UserInterface {
     @Field()
     id: string;