+++ packages/site/cms-site/src/iframebridge/IFrameBridge.tsx
@@ -8,6 +8,7 @@ export interface IFrameBridgeContext {
     hasBridge: boolean;
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     block?: any;
+    showOnlyVisible: boolean;
     selectedAdminRoute?: string;
     hoveredAdminRoute?: string;
     sendSelectComponent: (id: string) => void;
@@ -23,6 +24,7 @@ export interface IFrameBridgeContext {
 export const IFrameBridgeContext = React.createContext<IFrameBridgeContext>({
     hasBridge: false,
     showOutlines: false,
+    showOnlyVisible: false,
     sendSelectComponent: () => {
         // empty
     },
@@ -37,6 +39,7 @@ export const IFrameBridgeContext = React.createContext<IFrameBridgeContext>({
 
 export const IFrameBridgeProvider: React.FunctionComponent<{ children: React.ReactNode }> = ({ children }) => {
     const [block, setBlock] = React.useState<unknown | undefined>(undefined);
+    const [showOnlyVisible, setShowOnlyVisible] = React.useState<boolean>(false);
     const [selectedAdminRoute, setSelectedAdminRoute] = React.useState<string | undefined>(undefined);
     const [hoveredAdminRoute, setHoveredAdminRoute] = React.useState<string | undefined>(undefined);
     const [showOutlines, setShowOutlines] = React.useState<boolean>(false);
@@ -55,6 +58,9 @@ export const IFrameBridgeProvider: React.FunctionComponent<{ children: React.Rea
             switch (message.cometType) {
                 case AdminMessageType.Block:
                     setBlock(message.data.block);
+                    break;
+                case AdminMessageType.ShowOnlyVisible:
+                    setShowOnlyVisible(message.data.showOnlyVisible);
                     break;
                 case AdminMessageType.SelectComponent:
                     setSelectedAdminRoute(
@@ -109,6 +115,7 @@ export const IFrameBridgeProvider: React.FunctionComponent<{ children: React.Rea
                 showOutlines,
                 hasBridge: true,
                 block,
+                showOnlyVisible,
                 selectedAdminRoute,
                 hoveredAdminRoute,
                 sendSelectComponent: (adminRoute: string) => {